# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ChattyCathy API
  description: |
    ChattyCathy API Server with JWT authentication and RBAC.

    ## Authentication

    This API uses JWT (RS256) authentication with access and refresh tokens.

    - **Access tokens**: Short-lived (15 min), used in Authorization header
    - **Refresh tokens**: Long-lived (7 days), used to obtain new access tokens

    ### Authentication Methods

    - **Google OAuth**: Sign in with Google (recommended for users)
    - **Username/Password**: Traditional login (for demo/testing)

    ### For Web Clients
    Tokens are automatically set as HTTP-only cookies after login.

    ### For Mobile/API Clients
    Use the `Authorization: Bearer <token>` header with access tokens.

    ## Role-Based Access Control (RBAC)

    The API implements permission-based access control:

    ### Default Roles
    - **admin**: Full access to all resources
    - **user**: Basic read access (ping, news) - *automatically assigned to new users*
    - **editor**: Read and write access to content

    ### Default Permissions
    - `ping:read` - Access ping endpoint
    - `news:read`, `news:create`, `news:update`, `news:delete`
    - `users:read`, `users:create`, `users:update`, `users:delete`
    - `roles:read`, `roles:create`, `roles:update`, `roles:delete`

    ### Automatic Role Assignment
    New users registered via Google OAuth are automatically assigned the **user** role,
    which grants `ping:read` and `news:read` permissions.

...
servers:
...
paths:
...
  /auth/login:
    post:
      summary: Login (Demo)
      description: |
        Authenticate with email and password.

        **Note**: This is primarily for demo/testing purposes.
        For production, use Google OAuth.

        Returns access and refresh tokens in the response body.
        For web clients, tokens are also set as HTTP-only cookies.
      operationId: login
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: HTTP-only cookies for access_token and refresh_token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      summary: Refresh tokens
      description: |
        Get new access and refresh tokens using a valid refresh token.

        The refresh token can be provided via:
        - HTTP-only cookie (web clients)
        - Request body (`refresh_token` field)
        - Header (`X-Refresh-Token`)

        **Note**: Refresh tokens are rotated on each use. The old refresh token becomes invalid.
      operationId: refresh
      tags:
        - auth
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed successfully
          headers:
            Set-Cookie:
              description: New HTTP-only cookies for access_token and refresh_token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      summary: Logout
      description: |
        Logout the current session by revoking the refresh token.

        The refresh token can be provided via cookie, body, or header.
      operationId: logout
      tags:
        - auth
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout-all:
    post:
      summary: Logout all sessions
      description: Revoke all refresh tokens for the current user
      operationId: logoutAll
      tags:
        - auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutAllResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/sessions:
    get:
      summary: List active sessions
      description: Get a list of all active sessions for the current user
      operationId: listSessions
      tags:
        - auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of active sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: john
        password:
          type: string
          format: password
          example: password123
      required:
        - username
        - password
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: The refresh token (if not using cookies)
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token expiry in seconds
          example: 900
        refresh_expires_in:
          type: integer
          description: Refresh token expiry in seconds
          example: 604800
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
        - refresh_expires_in
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LogoutAllResponse:
      type: object
      properties:
        message:
          type: string
          example: all sessions revoked
        revoked_count:
          type: integer
          description: Number of sessions revoked
          example: 3
      required:
        - message
        - revoked_count
    SessionsResponse:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"
      required:
        - sessions
    Session:
      type: object
      properties:
        token_id:
          type: string
          description: Unique session identifier
        user_agent:
          type: string
          description: Browser/client user agent
        ip_address:
          type: string
          description: IP address of the session
        created_at:
          type: string
          format: date-time
          description: When the session was created
        expires_at:
          type: string
          format: date-time
          description: When the session expires
      required:
        - token_id
        - created_at
        - expires_at
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...